+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## Added
 - Support for [FileIcons](https://github.com/file-icons/icons). (#204)
+- Support for [Weather Icons](https://github.com/erikflowers/weather-icons). (#204)
 
 # [6.5.0] - 2024-09-14
 ## Changed